requirements.txt
google-generativeai
docx2pdf
streamlit
PyMuPDF
python-docx
pillow